add lava, buttons off after win or death, norm speed of map and spells, norm jump, ss debug, new lottery, new circle of tutor

+++ Assets/Scripts/spellsystem/DefSpellParent.cs
@@ -3,7 +3,7 @@ using System.Collections;
 
 abstract public class DefSpellParent : MonoBehaviour {
 
-	protected int[] defCount = new int[3];
+	protected float[] defCount = new float[3];
 	protected float[] defRestoreTime = new float[3];
 	protected float[] defRestoreSpeed = new float[3];
 	protected float[] defSpecial = new float[3];
